criação da rota para cadastro de produto e dropdown no Meu Perfil
