+++ src/app/components/cadastro-produto/cadastro-produto.component.spec.ts
@@ -1,0 +1,23 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { CadastroProdutoComponent } from './cadastro-produto.component';
+
+describe('CadastroProdutoComponent', () => {
+  let component: CadastroProdutoComponent;
+  let fixture: ComponentFixture<CadastroProdutoComponent>;
+
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      declarations: [CadastroProdutoComponent]
+    })
+    .compileComponents();
+
+    fixture = TestBed.createComponent(CadastroProdutoComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
