+++ src/app/components/cadastro-usuario/cadastro-usuario.component.css
@@ -68,6 +68,7 @@ main form .inputs-container input {
     color: #312e42;
     box-shadow: 0px 3px 3px rgba(0, 0, 0, 0.25);
 }
+
 main form .inputs-container input:focus {
     box-shadow: inset 1px 4px 7px rgba(0, 0, 0, 0.5);
 }
