+++ src/app/components/header/header.component.css
@@ -13,6 +13,14 @@
   box-shadow: 0 4px 6px rgba(0, 0, 0, 0.2);
 }
 
+.version {
+  position: absolute;
+  margin: 0;
+  color: gray;
+  left: 5px;
+  top: 3px;
+}
+
 .top-section {
   display: flex;
   width: 100%;
@@ -102,6 +110,47 @@
 }
 
 .navigation a:hover {
+  color: white;
+}
+
+.drop-div {
+  position: relative;
+  display: inline-block;
+  border-radius: 40%;
+}
+
+.drop-content {
+  display: none;
+  position: absolute;
+  background-color: #333;
+  min-width: 180px;
+  box-shadow: 0px 8px 16px 0px rgba(0,0,0,0.2);
+  z-index: 1;
+  border-radius: 4px;
+}
+
+.drop-content a {
+  color: orangered;
+  font-size: 1em;
+  padding: 12px 16px;
+  text-decoration: none;
+  display: block;
+  -webkit-transition: all 0.4s;
+}
+
+.drop-content a:hover {
+  background-color: #ddd;
+  color: black;
+  font-weight: 500;
+  transform: scale(1.05);
+  border-radius: 3px;
+}
+
+.drop-div:hover .drop-content {
+  display: block;
+}
+
+.drop-div:hover .MeuPerfil {
   color: white;
 }
