+++ src/app/components/header/header.component.html
@@ -1,4 +1,5 @@
 <div class="header">
+  <p class="version">Versão - 1.0</p>
   <div class="top-section d-flex align-items-center">
     <div class="logo">
       <img id="imagem" src="assets/img/logo.png" alt="img">
@@ -9,7 +10,7 @@
         <fa-icon [icon]="faSearch"></fa-icon>
       </button>
       <div>
-        <button routerLink="" class="new-post-button">
+        <button routerLink="/cadastro-produto" class="new-post-button">
           <fa-icon [icon]="faPlus"></fa-icon>
         </button>
       </div>
@@ -21,7 +22,14 @@
       <li class="me-3"><a id="Home" routerLink="">Home</a></li>
       <li class="me-3"><a id="DestaquesdaSemana" routerLink="">Destaques da Semana</a></li>
       <li><a id="Recentes" routerLink="">Recentes</a></li>
-      <li><a id="MeuPerfil" routerLink="/cadastro-usuario">Meu Perfil</a></li>
+      <li class="drop-div">
+        <a class="MeuPerfil" routerLink="/cadastro-usuario">Meu Perfil</a>
+        <div class="drop-content">
+          <a routerLink="/cadastro-usuario">Login/Cadastro</a>
+          <a routerLink="">Editar Publicação</a>
+          <a routerLink="">Configurações</a>
+        </div>
+      </li>
     </ul>
   </nav>
 </div>
