--- src/app/components/home/home.component.css
@@ -1,145 +1,0 @@
-/* Estilo geral */
-body {
-  font-family: Arial, sans-serif;
-  margin: 0;
-  padding: 0;
-  background-color: #f7f7f7;
-}
-
-/* Estilo da seção de formulário */
-#product-form {
-  max-width: 600px;
-  margin: 20px auto;
-  padding: 20px;
-  background-color: #fff;
-  border-radius: 8px;
-  box-shadow: 0 4px 10px rgba(0, 0, 0, 0.3);
-}
-
-#product-form h2 {
-  text-align: center;
-  color: #333;
-  margin: 50px 0 40px 0;
-}
-
-.h2Cadastrar {
-  font-family: Arial, sans-serif;
-  font-weight: 800;
-}
-
-#product-form label {
-  display: block;
-  margin-bottom: 5px;
-  color: #333;
-}
-
-#product-form input {
-  width: 100%;
-  padding: 10px;
-  margin-bottom: 15px;
-  border: 1px solid #ddd;
-  border-radius: 5px;
-}
-
-#product-form .btn-cadastrar {
-  width: 100%;
-  padding: 12px;
-  background-color: #ff4500;
-  color: #fff;
-  border: none;
-  border-radius: 5px;
-  cursor: pointer;
-  font-size: 16px;
-  font-weight: bold;
-  transition: background-color 0.3s;
-}
-
-#product-form .btn-cadastrar:hover {
-  background-color: #e63900;
-  transform: scale(1.03);
-  transition: 0.3s;
-}
-
-#product-form .btn-cadastrar:active {
-  transform: scale(1);
-  transition: 0.3s;
-}
-
-/* Estilo da lista de produtos */
-#product-container {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 20px;
-  justify-content: center;
-  padding: 60px;
-}
-
-.product-card {
-  width: 280px;
-  background-color: #fff;
-  border-radius: 8px;
-  box-shadow: 0 2px 5px rgba(0, 0, 0, 0.1);
-  overflow: hidden;
-  text-align: center;
-  padding: 15px;
-  transition: transform 0.3s;
-}
-
-.product-card:hover {
-  transform: translateY(-5px);
-}
-
-.product-card img {
-  width: 100%;
-  height: 180px;
-  object-fit: cover;
-  border-radius: 8px;
-}
-
-.product-card h3 {
-  font-size: 18px;
-  margin: 10px 0;
-  color: #333;
-}
-
-.product-card p {
-  font-size: 14px;
-  color: #555;
-}
-
-.error-message{
-  margin-left: 5px;
-  margin-top: 10px;
-  font-family: Georgia, 'Times New Roman', Times, serif;
-  color: #e63900;
-}
-#text{
-  font-family:Verdana, Geneva, Tahoma, sans-serif
-}
-/* Estilo dos ícones de ação */
-.icon-actions {
-  display: flex;
-  justify-content: center;
-  gap: 15px;
-  margin-top: 10px;
-}
-
-.icon-actions fa-icon {
-  font-size: 20px;
-  transition: transform 0.3s;
-}
-
-.icon-actions fa-icon:hover {
-  transform: scale(1.2);
-  color: #e63900;
-}
-.error {
-  color: red;
-  font-size: 0.8em;
-}
-
-.container{
- padding: 30px;
- margin-top: 100px;
-}
-
