+++ src/app/components/home/home.component.html
@@ -1,94 +1,2 @@
-<div class="container">
-  <section id="product-form">
-    <h2 class="h2Cadastrar">Cadastrar Novo Produto</h2>
-    <form id="new-product-form" (submit)="addProduct($event)">
-      <label id="text" for="product-image">Imagem (URL):</label>
-      <input
-        type="text"
-        id="product-image"
-        placeholder="URL da imagem"
-        [(ngModel)]="newProduct.image"
-        name="image"
-        required
-      />
-
-      <label id="text" for="promotion-link">Link da Promoção (URL):</label>
-      <input
-        type="text"
-        id="promotion-link"
-        placeholder="URL da promoção"
-        [(ngModel)]="newProduct.promotionLink"
-        name="promotionLink"
-        required
-      />
-
-      <label id="text" for="product-name">Nome do Produto:</label>
-      <input
-        type="text"
-        id="product-name"
-        placeholder="Nome do Produto"
-        [(ngModel)]="newProduct.name"
-        name="name"
-        required
-      />
-
-      <label id="text" for="product-price">Preço:</label>
-      <input
-        type="text"
-        id="product-price"
-        placeholder="Preço do Produto"
-        [(ngModel)]="newProduct.price"
-        name="price"
-        mask="separator.2"
-        prefix="R$ "
-        required
-      />
-
-      <label id="text" for="product-description">Descrição:</label>
-      <input
-        type="text"
-        id="product-description"
-        placeholder="Descrição do Produto"
-        [(ngModel)]="newProduct.description"
-        name="description"
-        required
-      />
-
-      <button type="submit" class="btn-cadastrar">
-        Cadastrar Produto
-      </button>
-      <div class="error-message">
-        <span *ngIf="isClicked && !isFormValid()">
-          Preencha todos os campos corretamente.
-        </span>
-      </div>
-    </form>
-  </section>
-
-  <section class="container" id="product-container">
-    <div
-      *ngFor="let product of products"
-      class="product-card"
-      (click)="openPromotionLink(product.promotionLink)"
-    >
-      <img [src]="product.image" alt="{{ product.name }}" />
-      <h3>{{ product.name }}</h3>
-      <h2>Preço: {{ formatPriceForDisplay(product.price) }}</h2>
-      <p>{{ product.description }}</p>
-
-      <!-- Ícones de like e dislike -->
-      <div class="icon-actions">
-        <fa-icon
-          [icon]="faThumbsUp"
-          style="color: #ff4500; cursor: pointer"
-          title="Like"
-        ></fa-icon>
-        <fa-icon
-          [icon]="faThumbsDown"
-          style="color: #ff4500; cursor: pointer"
-          title="Dislike"
-        ></fa-icon>
-      </div>
-    </div>
-  </section>
-</div>
+<br><br><br><br><br><br><br><br><br><br>
+<h1>HOME</h1>
